move audit log to statedb
